+++ src/commands/delete/index.js
@@ -1,0 +1,5 @@
+async function deleteItems(args) {
+	console.log(args);
+}
+
+module.exports = deleteItems;